Backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.1
stripe==8.1.0
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
